Modularize excursions page
Split Excursions page into per-section components under
src/components/excursions/*, lift static content and helpers into
src/data/excursionsPageContent.js, and reuse usePageMeta /
useRevealOnScroll hooks. Split src/styles/excursions.css into a manifest
of @imports under src/styles/excursions/* with cascade order preserved.
