+++ src/styles/excursions/anchor-offsets.css
@@ -1,0 +1,2 @@
+/* anchor offset for sticky filter strip */
+.exc-day, #day { scroll-margin-top: 120px; }
