+++ src/styles/excursions/faq.css
@@ -1,0 +1,30 @@
+/* ============ FAQ ============ */
+.exc-faq { padding: clamp(4rem, 7vw, 6rem) clamp(1.25rem, 5vw, 6rem); }
+.exc-faq__head { text-align: center; max-width: 600px; margin: 0 auto 3rem; }
+.exc-faq__list { max-width: 820px; margin: 0 auto; display: flex; flex-direction: column; gap: .5rem; }
+.exc-faq__item {
+  border: 1px solid color-mix(in srgb, currentColor 14%, transparent);
+  border-radius: 10px;
+  background: var(--surface, #FFFCF5);
+}
+.exc-faq__item summary {
+  list-style: none;
+  padding: 1.25rem 1.5rem;
+  cursor: pointer;
+  font-weight: 500;
+  font-size: 1rem;
+  color: var(--text, #1a4d6e);
+  display: flex; justify-content: space-between; align-items: center; gap: 1rem;
+}
+.exc-faq__item summary::-webkit-details-marker { display: none; }
+.exc-faq__item summary::after {
+  content: '+';
+  font-size: 1.5rem;
+  color: #FF6F61;
+  font-weight: 300;
+  flex-shrink: 0;
+  transition: transform .25s;
+}
+.exc-faq__item[open] summary::after { content: '−'; }
+.exc-faq__body { padding: 0 1.5rem 1.4rem; font-size: .95rem; line-height: 1.65; opacity: .85; }
+.exc-faq__body a { color: #FF6F61; }
