+++ src/styles/excursions/grid.css
@@ -1,0 +1,262 @@
+/* ============ INDEX CARD GRID ============ */
+.exc-grid-wrap { padding: clamp(3rem, 6vw, 5rem) clamp(1.25rem, 5vw, 4rem) clamp(3rem, 6vw, 5rem); }
+.exc-list__head {
+  max-width: 760px;
+  margin: 0 auto clamp(1.5rem, 2.8vw, 2.25rem);
+  text-align: center;
+}
+.exc-list__head .section-title {
+  margin-bottom: clamp(.75rem, 1.4vw, 1rem);
+}
+.exc-list__head .section-lead {
+  max-width: 680px;
+}
+.exc-grid-wrap > .exc-filter {
+  max-width: 1240px;
+  margin: 0 auto clamp(1.2rem, 2vw, 1.75rem);
+}
+.exc-grid {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  gap: clamp(1.25rem, 2vw, 1.75rem);
+  max-width: 1240px;
+  margin: 0 auto;
+}
+.exc-grid .exc-card {
+  flex: 1 1 310px;
+  max-width: 380px;
+}
+
+@media (max-width: 720px) {
+  .exc-grid-wrap {
+    padding: 3rem 1rem 3.25rem;
+  }
+
+  .exc-list__head {
+    margin-bottom: 1.15rem;
+  }
+
+  .exc-grid-wrap > .exc-filter,
+  .exc-filter {
+    position: static;
+    max-width: none;
+    margin: 0 0 1.5rem;
+    padding: .2rem 0 0;
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    gap: .45rem;
+    overflow: visible;
+  }
+
+  .exc-filter__btn {
+    flex: 0 0 auto;
+    min-height: 38px;
+    padding: .45rem .75rem;
+    border-color: color-mix(in srgb, var(--text-soft) 62%, transparent);
+    background: var(--surface);
+    color: var(--text-soft);
+    opacity: 1;
+    font-size: .82rem;
+    line-height: 1.1;
+    white-space: nowrap;
+  }
+
+  .exc-filter__btn.is-active {
+    color: #fff;
+  }
+
+  .exc-grid {
+    gap: 1rem;
+  }
+
+  .exc-grid .exc-card {
+    flex-basis: 100%;
+    max-width: 100%;
+  }
+}
+.exc-grid__empty {
+  text-align: center;
+  padding: 3rem 1rem;
+  color: var(--text-muted);
+  font-size: .95rem;
+}
+.exc-grid__actions {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-wrap: wrap;
+  gap: .9rem 1.15rem;
+  max-width: 1240px;
+  margin: clamp(1.5rem, 3vw, 2.25rem) auto 0;
+  color: color-mix(in srgb, currentColor 62%, transparent);
+  font-size: .92rem;
+  font-weight: 700;
+}
+.exc-grid__actions span {
+  letter-spacing: .08em;
+  text-transform: uppercase;
+}
+
+.exc-card {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  background: var(--surface, #FFFCF5);
+  border: 1px solid color-mix(in srgb, currentColor 14%, transparent);
+  border-radius: 14px;
+  overflow: hidden;
+  text-decoration: none;
+  color: var(--text, #1a4d6e);
+  box-shadow: 0 14px 32px -28px rgba(0,0,0,.45);
+  transition: transform .35s ease, box-shadow .35s ease, border-color .35s ease;
+}
+.exc-card:hover {
+  transform: translateY(-4px);
+  box-shadow: 0 28px 50px -28px rgba(8,30,46,.45);
+  border-color: rgba(255,111,97,.35);
+}
+
+.exc-card__img {
+  position: relative;
+  aspect-ratio: 4 / 3;
+  overflow: hidden;
+}
+.exc-card__img img {
+  width: 100%; height: 100%;
+  object-fit: cover;
+  transition: transform 1.1s ease;
+}
+.exc-card:hover .exc-card__img img { transform: scale(1.05); }
+
+/* placeholder when no real photo yet */
+.exc-card__img--placeholder,
+.exc-block__img--placeholder {
+  background: #071f2c;
+}
+
+.exc-card__cat {
+  position: absolute;
+  top: .85rem;
+  left: .85rem;
+  font-size: 9.5px;
+  letter-spacing: .18em;
+  text-transform: uppercase;
+  background: rgba(255,255,255,.95);
+  color: #1a4d6e;
+  padding: .35rem .65rem;
+  border-radius: 999px;
+  font-weight: 700;
+  backdrop-filter: blur(6px);
+}
+.exc-card__cat[data-cat="water-snorkeling"] { background: #c9e7f0; color: #0a3a55; }
+.exc-card__cat[data-cat="culture-heritage"] { background: #f5d9c0; color: #6e3010; }
+.exc-card__cat[data-cat="nature-wildlife"] { background: #d2e2c1; color: #2f4922; }
+.exc-card__cat[data-cat="adventure-sports"] { background: #fed7aa; color: #7c2d12; }
+.exc-card__cat[data-cat="wellness-luxury"] { background: #ddd6fe; color: #4c1d95; }
+.exc-card__cat[data-cat="festivals-seasonal"] { background: #fde68a; color: #78350f; }
+
+.exc-card__season {
+  position: absolute;
+  bottom: .85rem;
+  left: .85rem;
+  font-size: 9.5px;
+  letter-spacing: .14em;
+  text-transform: uppercase;
+  background: #FF6F61;
+  color: #fff;
+  padding: .35rem .65rem;
+  border-radius: 999px;
+  font-weight: 700;
+  box-shadow: 0 4px 12px -4px rgba(255,111,97,.7);
+}
+
+.exc-card__body {
+  padding: 1.1rem 1.15rem 1.25rem;
+  display: flex;
+  flex-direction: column;
+  gap: .55rem;
+  flex: 1;
+}
+.exc-card__eyebrow {
+  font-size: 10px;
+  letter-spacing: .18em;
+  text-transform: uppercase;
+  color: #FF6F61;
+  font-weight: 700;
+}
+.exc-card__title {
+  font-family: var(--dp-font-display, 'Playfair Display', Georgia, serif);
+  font-size: 1.3rem;
+  line-height: 1.2;
+  font-weight: 500;
+  margin: 0;
+  letter-spacing: -.005em;
+  color: var(--text, #1a4d6e);
+}
+.exc-card__desc {
+  font-size: .88rem;
+  line-height: 1.55;
+  color: currentColor;
+  opacity: .8;
+  margin: 0;
+  display: -webkit-box;
+  -webkit-line-clamp: 3;
+  line-clamp: 3;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+}
+.exc-card__meta {
+  display: flex;
+  flex-wrap: wrap;
+  gap: .85rem;
+  font-size: 11.5px;
+  color: color-mix(in srgb, currentColor 75%, transparent);
+  margin-top: .15rem;
+}
+.exc-card__meta span {
+  display: inline-flex;
+  align-items: center;
+  gap: .35rem;
+}
+.exc-card__meta svg { color: #FF6F61; }
+
+.exc-card__foot {
+  margin-top: auto;
+  padding-top: .85rem;
+  border-top: 1px solid color-mix(in srgb, currentColor 14%, transparent);
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: .5rem;
+}
+.exc-card__price {
+  font-size: 12.5px;
+  color: color-mix(in srgb, currentColor 75%, transparent);
+}
+.exc-card__price strong {
+  font-family: var(--dp-font-display, 'Playfair Display', Georgia, serif);
+  font-size: 1.05rem;
+  color: #FF6F61;
+  font-weight: 500;
+  margin: 0 .15rem 0 .25rem;
+}
+.exc-card__price--tbd {
+  font-size: 11px;
+  letter-spacing: .12em;
+  text-transform: uppercase;
+  color: color-mix(in srgb, currentColor 60%, transparent);
+}
+.exc-card__cta {
+  font-size: 12px;
+  letter-spacing: .12em;
+  text-transform: uppercase;
+  color: #FF6F61;
+  font-weight: 700;
+}
+
+@media (max-width: 540px) {
+  .exc-grid { grid-template-columns: 1fr; gap: 1rem; }
+  .exc-card__img { aspect-ratio: 16 / 11; }
+}
