+++ src/styles/excursions/hero.css
@@ -1,0 +1,146 @@
+/* ============ HERO (mirrors .saf-hero pattern) ============ */
+.exc-hero {
+  position: relative;
+  min-height: 100svh;
+  display: flex; align-items: center; justify-content: center;
+  text-align: center;
+  padding: clamp(4.75rem, 8vw, 6.5rem) clamp(1.25rem, 5vw, 6rem) clamp(2.25rem, 5vh, 3.5rem);
+  color: #F5F1E8;
+  overflow: hidden;
+  isolation: isolate;
+}
+.exc-hero__bg { position: absolute; inset: 0; z-index: 0; }
+.exc-hero__bg img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  transform: scale(1.05);
+  animation: heroImageReveal 1.4s var(--dp-ease-out) both;
+}
+.exc-hero__bg::after {
+  content: '';
+  position: absolute; inset: 0;
+  background:
+    linear-gradient(120deg, rgba(8,30,46,.85) 0%, rgba(8,30,46,.55) 50%, rgba(8,30,46,.25) 100%),
+    linear-gradient(180deg, transparent 50%, rgba(8,30,46,.7) 100%);
+}
+.exc-hero__inner {
+  position: relative;
+  z-index: 1;
+  width: min(780px, 100%);
+  max-width: 780px;
+  margin: 0 auto;
+  display: grid;
+  justify-items: center;
+  gap: clamp(.75rem, 1.35vh, 1.3rem);
+}
+.exc-hero__eyebrow {
+  display: inline-flex; align-items: center; gap: .5rem;
+  font-family: var(--dp-font-sans);
+  font-size: 11px; letter-spacing: .18em; text-transform: uppercase;
+  background: rgba(255,255,255,.12);
+  backdrop-filter: blur(10px);
+  border: 1px solid rgba(255,255,255,.18);
+  padding: .5rem 1rem;
+  border-radius: 999px;
+  margin-bottom: 0;
+  animation: fadeInUp .9s var(--dp-ease-out) both;
+}
+.exc-hero__eyebrow::before { content: ''; width: 6px; height: 6px; border-radius: 50%; background: var(--dp-accent, #FF6F61); box-shadow: 0 0 10px var(--dp-accent, #FF6F61); }
+.exc-hero__title {
+  font-family: var(--dp-font-display);
+  font-size: clamp(2.75rem, 6.5vw, 5.25rem);
+  line-height: 1.02;
+  font-weight: 500;
+  margin: 0;
+  letter-spacing: -.015em;
+  color: #fff;
+  animation: fadeInUp 1s var(--dp-ease-out) both;
+  animation-delay: .1s;
+}
+.exc-hero__title em { font-style: italic; color: var(--dp-accent, #FF6F61); font-weight: 400; }
+.exc-hero__lead {
+  font-family: var(--dp-font-display);
+  font-size: clamp(1rem, 1.4vw, 1.2rem);
+  line-height: 1.65;
+  color: rgba(245,241,232,.88);
+  max-width: 640px;
+  margin: 0 auto;
+  animation: fadeInUp 1s var(--dp-ease-out) both;
+  animation-delay: .3s;
+}
+.exc-hero__tags {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  gap: .5rem;
+  margin-top: .1rem;
+  animation: fadeInUp 1s var(--dp-ease-out) both;
+  animation-delay: .45s;
+}
+.exc-hero__tags span {
+  display: inline-flex;
+  align-items: center;
+  gap: .4rem;
+  padding: .4rem .75rem;
+  border-radius: 999px;
+  border: 1px solid rgba(255,255,255,.28);
+  background: rgba(255,255,255,.1);
+  color: rgba(245,241,232,.96);
+  font-family: var(--dp-font-sans);
+  font-size: 10px;
+  letter-spacing: .12em;
+  text-transform: uppercase;
+  font-weight: 600;
+  backdrop-filter: blur(8px);
+}
+.exc-hero__row {
+  display: flex; gap: .75rem; flex-wrap: wrap; justify-content: center;
+  margin-bottom: clamp(.25rem, .7vh, .65rem);
+  animation: fadeInUp 1s var(--dp-ease-out) both;
+  animation-delay: .5s;
+}
+.exc-hero__meta {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: 1.25rem;
+  max-width: 640px;
+  width: 100%;
+  padding-top: clamp(.9rem, 1.7vh, 1.3rem);
+  border-top: 1px solid rgba(255,255,255,.18);
+  animation: fadeInUp 1s var(--dp-ease-out) both;
+  animation-delay: .7s;
+}
+.exc-hero__meta div {
+  display: flex; flex-direction: column; gap: 4px;
+  padding: 0; border: 0; background: transparent; backdrop-filter: none; border-radius: 0;
+}
+.exc-hero__meta strong {
+  font-family: var(--dp-font-display);
+  font-size: clamp(1.5rem, 2.4vw, 2.2rem);
+  font-weight: 500;
+  line-height: 1;
+}
+.exc-hero__meta span {
+  font-family: var(--dp-font-sans);
+  font-size: 11px; letter-spacing: .12em; text-transform: uppercase; opacity: .65;
+  margin-top: 0;
+}
+@media (max-width: 720px) {
+  .exc-hero__meta { grid-template-columns: repeat(2, 1fr); gap: 1.25rem 2rem; }
+}
+@media (min-width: 721px) and (max-height: 860px) {
+  .exc-hero {
+    padding-top: clamp(4rem, 7vw, 5rem);
+    padding-bottom: 1.75rem;
+  }
+  .exc-hero__title { font-size: clamp(2.35rem, 5.1vw, 4.2rem); }
+  .exc-hero__lead { font-size: clamp(0.95rem, 1.15vw, 1.08rem); line-height: 1.52; }
+  .exc-hero__meta strong { font-size: clamp(1.3rem, 2vw, 1.85rem); }
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .exc-hero__bg img {
+    animation: none;
+  }
+}
