+++ src/styles/excursions/pairings.css
@@ -1,0 +1,46 @@
+/* ============ PAIRINGS ============ */
+.exc-pair {
+  background: linear-gradient(180deg, color-mix(in srgb, currentColor 4%, transparent) 0%, color-mix(in srgb, currentColor 9%, transparent) 100%);
+  padding: clamp(4rem, 7vw, 6rem) clamp(1.25rem, 5vw, 6rem);
+}
+.exc-pair__head { text-align: center; max-width: 600px; margin: 0 auto 3rem; }
+.exc-pair__grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+  gap: 1.5rem;
+  max-width: 1240px;
+  margin: 0 auto;
+}
+.exc-pair__card {
+  background: var(--surface, #FFFCF5);
+  border-radius: 12px;
+  padding: 1.75rem;
+  border: 1px solid color-mix(in srgb, currentColor 12%, transparent);
+  display: flex; flex-direction: column; gap: 1rem;
+  color: var(--text, #1a4d6e);
+  text-decoration: none;
+  box-shadow: 0 14px 32px -28px rgba(0,0,0,.4);
+  transition: transform .3s ease, border-color .3s ease, box-shadow .3s ease;
+}
+.exc-pair__card:hover {
+  transform: translateY(-4px);
+  border-color: color-mix(in srgb, var(--dp-accent) 50%, currentColor 18%);
+  box-shadow: 0 22px 44px -32px rgba(0,0,0,.52);
+}
+.exc-pair__card:focus-visible {
+  outline: 2px solid var(--dp-accent);
+  outline-offset: 4px;
+}
+.exc-pair__combo { display: flex; align-items: center; gap: .65rem; flex-wrap: wrap; font-size: 11px; letter-spacing: .14em; text-transform: uppercase; color: #FF6F61; font-weight: 600; }
+.exc-pair__combo span { padding: .3rem .6rem; background: rgba(255,111,97,.1); border-radius: 4px; }
+.exc-pair__combo svg { color: rgba(255,111,97,.6); }
+.exc-pair__title {
+  font-family: var(--dp-font-display, 'Playfair Display', Georgia, serif);
+  font-size: 1.35rem; font-weight: 500;
+  line-height: 1.2;
+  margin: 0;
+  color: var(--text, #1a4d6e);
+}
+.exc-pair__desc { font-size: .92rem; line-height: 1.6; opacity: .8; margin: 0; }
+.exc-pair__foot { display: flex; justify-content: space-between; align-items: center; padding-top: 1rem; border-top: 1px solid color-mix(in srgb, currentColor 12%, transparent); font-size: 12px; }
+.exc-pair__foot strong { color: #FF6F61; font-family: var(--dp-font-display, 'Playfair Display', Georgia, serif); font-size: 1.1rem; }
