+++ src/styles/excursions/reveal.css
@@ -1,0 +1,3 @@
+/* ============ REVEAL ============ */
+.excursions-page .reveal { opacity: 0; transform: translateY(24px); transition: opacity .8s ease, transform .8s ease; }
+.excursions-page .reveal.is-visible { opacity: 1; transform: none; }
